feat: add release.yaml specific for batch changes (#291)
## Description
Since we're dependent on batch-changes for pr creation now (given it
doesn't clone the .git files so we can't create PRs from execution
phase) we have to modify the release.yaml. Instead of breaking existing
workflows, we're creating a new release.yaml here eventually to be
deprecated. I tested this yesterday on the docker repo, and it worked
fine, so I'm replicating that here.

It's a direct copy of the release.yaml, except since it's in a subdir,
it needs to change the directory sg ops update-images runs from. I also
had to remove references to git and gh

---

## Checklist

<!--
Kubernetes, both Kustomize and Helm, and Docker Compose MUST be kept in
sync.
You should not merge a change here without a corresponding change in the
other repositories,
unless it is specific to this deployment type. If uneeded, add link or
explanation of why it is not needed here.
-->

- [ ] Update
[CHANGELOG.md](https://github.com/sourcegraph/deploy-sourcegraph-k8s/blob/main/CHANGELOG.md)
- [ ] Update [K8s Upgrade
notes](https://github.com/sourcegraph/sourcegraph/blob/main/doc/admin/updates/kubernetes.md)
- [ ] Kustomiz-specific changes
- [x] Update sister repository:
[deploy-sourcegraph-helm](https://github.com/sourcegraph/deploy-sourcegraph-helm)
- [x] Update sister repository:
[deploy-sourcegraph-docker](https://github.com/sourcegraph/deploy-sourcegraph-docker)
- [ ] Verify all images have a valid tag and SHA256 sum

## Test plan
run batch change temporal workflow

<!--
As part of SOC2/GN-104 and SOC2/GN-105 requirements, all pull requests
are REQUIRED to
provide a "test plan". A test plan is a loose explanation of what you
have done or
implemented to test this, as outlined in our Testing principles and
guidelines:

https://docs.sourcegraph.com/dev/background-information/testing_principles
  Write your test plan here after the "Test plan" header.
-->